00_GCP_Projects/04_GCP_Document_Classification/requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
deprecation==2.1.0
Flask==3.0.3
functions-framework==3.8.1
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==3.0.2
packaging==24.1
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.0.5
win32-setctime==1.1.0
